requirement.txt
omegaconf==2.3.0
scikit-learn==1.6.1
pytorch-lightning==2.5.0.post0
timm==1.0.14
nltk==3.9.1
lmdb==1.6.2
Pillow==11.1.0
imgaug==0.4.0
hydra-core==1.3.2
fvcore==0.1.5.post20220512
ray[tune]==2.42.1
ax-platform==0.5.0
PyYAML==6.0.2
tqdm==4.67.1
opencv-python==4.11.0.86
opendatasets==0.1.22
transformers==4.48.3
seaborn==0.13.2
ftfy==6.3.1
torch==2.6.0
torchvision==0.21.0
numpy==2.2.2
scikit-image==0.25.1
scipy==1.15.1
python-Levenshtein==0.26.1
pandas==2.2.3
python-dotenv==1.0.1
sentence-transformers==3.4.1
openai==1.62.0
qdrant-client==1.13.2
uvicorn==0.34.0
fastapi==0.115.8
fuzzywuzzy==0.18.0